// WebSites/controller/src/types/Content.ts
export type ContentKey = string;
export type ContentHash = string;
export type UrlString = string;

export type SimulatorSharedContent = {
  contentKey?: ContentKey;
  contentHash?: ContentHash;
  contentIndexUrl?: UrlString;
  assetsBaseUrl?: UrlString;
  unityMetaKey?: string;
  unityMetaHash?: string;
  unityMetaUrl?: UrlString;
};